requirements.txt
# ─────────────────────────────────────────────
# Fashion Reuse Studio — Python Requirements
# ─────────────────────────────────────────────

# ── Core Deep Learning ──────────────────────
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# ── Diffusers Ecosystem ──────────────────────
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
peft>=0.9.0
safetensors>=0.4.2
tokenizers>=0.19.0

# ── ControlNet / IP-Adapter helpers ─────────
controlnet-aux>=0.0.7
ip-adapter @ git+https://github.com/tencent-ailab/IP-Adapter.git

# ── Attention / Memory Optimization ─────────
xformers>=0.0.23
triton>=2.1.0

# ── Image Processing ─────────────────────────
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0
albumentations>=1.3.1
imageio>=2.31.0

# ── CLIP / Evaluation ───────────────────────
open-clip-torch>=2.24.0
lpips>=0.1.4
pytorch-fid>=0.3.0
torchmetrics[image]>=1.3.0    # FID, LPIPS via torchmetrics API
git+https://github.com/openai/CLIP.git  # OpenAI CLIP for benchmark scoring

# ── Local Deployment / Quantization ─────────
# bitsandbytes already listed under GPU/Performance section
# optimum[onnxruntime-gpu]>=1.17.0  # optional: ONNX export (uncomment if needed)

# ── Segmentation (SAM / GrabCut) ────────────
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git

# ── Data & Utilities ────────────────────────
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
tqdm>=4.66.0
pyyaml>=6.0.1
omegaconf>=2.3.0
einops>=0.7.0
matplotlib>=3.7.0
seaborn>=0.12.0

# ── API Server ───────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
aiofiles>=23.2.1
httpx>=0.27.0

# ── LLM for DIY Guide ───────────────────────
openai>=1.14.0
anthropic>=0.21.0
langchain>=0.1.12
langchain-openai>=0.1.0

# ── Logging / Experiment Tracking ───────────
wandb>=0.16.0
tensorboard>=2.15.0
loguru>=0.7.2

# ── GPU / Performance ───────────────────────
bitsandbytes>=0.43.0
flash-attn>=2.5.0  # optional, comment out if build fails

# ── Testing ──────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
